test(flashattention): split monolithic kernel test into mha/gqa/misc
The single test_flashattention.py (26 test methods, ~15 min as one TestFile)
was the serial floor of unit-test-tpu-v6e-1. Split the shared setUp/run_test
workhorse into flashattention_common.AttentionTestBase and partition the 26
methods into three files (mha=10, gqa=10, misc=6) so run_suite.py can spread
them across shards instead of running them as one indivisible block.

Logic is moved verbatim; no test behavior changes.
