+++ python/sgl_jax/test/flashattention_common.py
@@ -1,5 +1,3 @@
-import unittest
-
 import jax
 import jax.numpy as jnp
 import numpy as np
@@ -356,7 +354,7 @@ def create_test_data(
     return fb, current_kv_cache, q, k, v
 
 
-class TestAttention(CustomTestCase):
+class AttentionTestBase(CustomTestCase):
     """Test cases for the Attention layer."""
 
     def setUp(self):
@@ -542,547 +540,3 @@ class TestAttention(CustomTestCase):
             are_close,
             f"JAX output and expected output are not close, max diff: {max_diff}",
         )
-
-    def test_mha_prefill_accuracy_page_size_1(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (1, 128),
-            (125, 125),
-            (1024, 1024),
-            (123, 522),
-            (1, 511),
-            (512, 1024),
-        ]
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-        )
-
-    def test_mha_decode_accuracy_page_size_1(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (1, 119),
-            (1, 127),
-            (1, 128),
-            (1, 129),
-            (1, 133),
-            (1, 1001),
-            (1, 1023),
-            (1, 1024),
-            (1, 1025),
-        ]
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-        )
-
-    def test_mha_prefill_accuracy_page_size_8(self):
-        """
-        Test JAX attention accuracy against PyTorch reference
-        This test case will failed when batch size > 2, the second batch tokens will has wrong value, the first and third batch tokens are correct.
-        """
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (5, 17),
-            (5, 33),
-            (5, 5),
-        ]
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 8, jnp.bfloat16),
-        )
-
-    def test_mha_decode_accuracy_page_size_8(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (1, 17),
-            (1, 6),
-            (1, 5),
-        ]
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 8, jnp.bfloat16),
-        )
-
-    def test_mha_prefill_accuracy_page_size_64(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (1, 128),
-            (3, 20),
-            (64, 64),
-            (20, 20),
-            (125, 125),
-            (1024, 1024),
-            (123, 522),
-            (1, 511),
-        ]
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16),
-        )
-
-    def test_mha_decode_accuracy_page_size_64(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        lens = [
-            (1, 20),
-            (1, 64),
-            (1, 30),
-            (1, 129),
-            (1, 133),
-            (1, 256),
-            (1, 1001),
-            (1, 1024),
-            (1, 1025),
-        ]
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16),
-        )
-
-    def test_gqa_prefill_accuracy_page_size_64(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 128),
-            (3, 20),
-            (64, 64),
-            (20, 20),
-            (125, 125),
-            (1024, 1024),
-            (123, 522),
-            (1, 511),
-        ]
-        self.run_test("prefill", lens, (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16))
-
-    def test_gqa_decode_accuracy_page_size_64(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 119),
-            (1, 127),
-            (1, 128),
-            (1, 129),
-            (1, 133),
-            (1, 1001),
-            (1, 1023),
-            (1, 1024),
-            (1, 1025),
-        ]
-
-        self.run_test("decode", lens, (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16))
-
-    def test_sliding_window_and_soft_cap_prefill_accuracy(self):
-        """Test combined sliding window and soft cap attention accuracy in prefill mode"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 256
-        sliding_window_size = 512
-        logit_cap = 20.0
-
-        lens = [
-            (1, 128),
-            (64, 64),
-            (128, 256),
-            (100, 300),
-            (1, 400),
-        ]
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            sliding_window=sliding_window_size,
-            logit_cap=logit_cap,
-        )
-
-    def test_sliding_window_and_soft_cap_decode_accuracy(self):
-        """Test combined sliding window and soft cap attention accuracy in decode mode"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-        sliding_window_size = 512
-        logit_cap = 20.0
-
-        lens = [
-            (1, 256),
-            (1, 400),
-            (1, 512),
-            (1, 1024),
-        ]
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            sliding_window=sliding_window_size,
-            logit_cap=logit_cap,
-        )
-
-    def test_gqa_prefill_accuracy_page_size_64_temperature(self):
-        """Test JAX attention accuracy against PyTorch reference
-        Testcase (1024, 1024) fails on token 607, possible precision issue?
-        Token 607: max_diff=0.023438, jax_mean=-0.011597, expected_mean=-0.011597, jax_std=0.048096, expected_std=0.047607
-        """
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 128),
-            (3, 20),
-            (64, 64),
-            (20, 20),
-            (125, 125),
-            (123, 522),
-            (1, 511),
-            (1024, 1024),
-        ]
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16),
-            xai_temperature_len=512,
-        )
-
-    def test_gqa_decode_accuracy_page_size_64_temperature(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 119),
-            (1, 127),
-            (1, 128),
-            (1, 129),
-            (1, 133),
-            (1, 1001),
-            (1, 1023),
-            (1, 1024),
-            (1, 1025),
-        ]
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 64, jnp.bfloat16),
-            xai_temperature_len=512,
-        )
-
-    def test_gqa_prefill_accuracy_page_size_1_temperature(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 128),
-            (3, 20),
-            (64, 64),
-            (20, 20),
-            (125, 125),
-            (1024, 1024),
-            (123, 522),
-            (1, 511),
-        ]
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            xai_temperature_len=512,
-        )
-
-    def test_gqa_decode_accuracy_page_size_1_temperature(self):
-        """Test JAX attention accuracy against native fa"""
-        # Parameters
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [
-            (1, 119),
-            (1, 127),
-            (1, 128),
-            (1, 129),
-            (1, 133),
-            (1, 1001),
-            (1, 1023),
-            (1, 1024),
-            (1, 1025),
-        ]
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            xai_temperature_len=512,
-        )
-
-    def test_mha_prefill_with_custom_mask(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 8
-        num_kv_heads = [8, 4]
-        head_dim = 128
-        lens = [(32, 32), (42, 66), (128, 256)]
-        page_size = [
-            1,
-        ]
-        causal_mask = False
-        for size in page_size:
-            for num_kv_head in num_kv_heads:
-                self.run_test(
-                    "prefill",
-                    lens,
-                    (num_heads, head_dim, num_kv_head, size, jnp.bfloat16, causal_mask),
-                )
-
-    def test_mha_decode_with_custom_mask(self):
-        pass
-
-    def test_gqa_prefill_with_custom_mask(self):
-        """Test JAX attention accuracy against PyTorch reference"""
-        # Parameters
-        num_heads = 128
-        num_kv_heads = 8
-        head_dim = 128
-        lens = [(32, 32), (42, 66), (128, 256)]
-        page_size = [
-            16,
-        ]
-        causal_mask = False
-        for size in page_size:
-            self.run_test(
-                "prefill",
-                lens,
-                (num_heads, head_dim, num_kv_heads, size, jnp.bfloat16, causal_mask),
-            )
-
-    def test_gqa_decode_with_custom_mask(self):
-        pass
-
-    def test_attention_sink_decode_accuracy(self):
-        """Test attention sink accuracy in decode mode with per-head sink logits"""
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-
-        lens = [
-            (1, 256),
-            (1, 512),
-            (1, 1024),
-        ]
-
-        # Per-head sink logits
-        rng = np.random.RandomState(123)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            attention_sink=attention_sink,
-        )
-
-    def test_attention_sink_prefill_accuracy(self):
-        """Test attention sink accuracy in prefill mode with per-head sink logits"""
-        num_heads = 32
-        num_kv_heads = 8
-        head_dim = 128
-
-        lens = [
-            (1, 128),
-            (64, 64),
-            (128, 256),
-        ]
-
-        # Per-head sink logits
-        rng = np.random.RandomState(456)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            attention_sink=attention_sink,
-        )
-
-    def test_mha_attention_sink_decode_accuracy(self):
-        """Test attention sink accuracy in decode mode with MHA (num_heads == num_kv_heads)"""
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-
-        lens = [
-            (1, 256),
-            (1, 512),
-            (1, 1024),
-        ]
-
-        rng = np.random.RandomState(789)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-    def test_mha_attention_sink_prefill_accuracy(self):
-        """Test attention sink accuracy in prefill mode with MHA (num_heads == num_kv_heads)"""
-        num_heads = 32
-        num_kv_heads = 32
-        head_dim = 128
-
-        lens = [
-            (1, 128),
-            (64, 64),
-            (128, 256),
-        ]
-
-        rng = np.random.RandomState(101)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-    def test_gqa_4q1kv_attention_sink_decode_accuracy(self):
-        """Test attention sink with GQA (4 q_heads, 1 kv_head) in decode mode"""
-        num_heads = 4
-        num_kv_heads = 1
-        head_dim = 128
-
-        lens = [
-            (1, 256),
-            (1, 512),
-            (1, 1024),
-        ]
-
-        rng = np.random.RandomState(202)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-    def test_gqa_4q1kv_attention_sink_prefill_accuracy(self):
-        """Test attention sink with GQA (4 q_heads, 1 kv_head) in prefill mode"""
-        num_heads = 4
-        num_kv_heads = 1
-        head_dim = 128
-
-        lens = [
-            (1, 128),
-            (64, 64),
-            (128, 256),
-        ]
-
-        rng = np.random.RandomState(303)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-    def test_single_head_attention_sink_decode_accuracy(self):
-        """Test attention sink with single head MHA (1 q_head, 1 kv_head) in decode mode"""
-        num_heads = 1
-        num_kv_heads = 1
-        head_dim = 128
-
-        lens = [
-            (1, 256),
-            (1, 512),
-            (1, 1024),
-        ]
-
-        rng = np.random.RandomState(404)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "decode",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-    def test_single_head_attention_sink_prefill_accuracy(self):
-        """Test attention sink with single head MHA (1 q_head, 1 kv_head) in prefill mode"""
-        num_heads = 1
-        num_kv_heads = 1
-        head_dim = 128
-
-        lens = [
-            (1, 128),
-            (64, 64),
-            (128, 256),
-        ]
-
-        rng = np.random.RandomState(505)
-        attention_sink = jnp.array(rng.randn(num_heads).astype(np.float32))
-
-        self.run_test(
-            "prefill",
-            lens,
-            (num_heads, head_dim, num_kv_heads, 1, jnp.bfloat16),
-            max_total_token_size=200000,
-            attention_sink=attention_sink,
-        )
-
-
-if __name__ == "__main__":
-    unittest.main()
